+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/users/RestUserGenerator.kt
@@ -1,16 +1,21 @@
 package com.atlassian.performance.tools.virtualusers.api.users
 
+import com.atlassian.data.utils.UserNameUtils
 import com.atlassian.performance.tools.jiraactions.api.memories.User
+import com.atlassian.performance.tools.jvmtasks.api.ExponentialBackoff
+import com.atlassian.performance.tools.jvmtasks.api.IdempotentAction
 import com.atlassian.performance.tools.jvmtasks.api.TaskTimer.time
 import com.atlassian.performance.tools.virtualusers.api.VirtualUserOptions
 import okhttp3.*
+import org.apache.http.entity.ContentType.APPLICATION_JSON
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
+import java.net.HttpURLConnection
 import java.time.Duration
-import java.util.*
 import java.util.concurrent.TimeUnit
 
 class RestUserGenerator(readTimeout: Duration) : UserGenerator {
+    private val nameGenerator = UserNameGenerator()
 
     private val logger: Logger = LogManager.getLogger(this::class.java)
     private val httpClient = OkHttpClient.Builder()
@@ -23,23 +28,26 @@ class RestUserGenerator(readTimeout: Duration) : UserGenerator {
         options: VirtualUserOptions
     ): User {
         val target = options.target
-        val uuid = UUID.randomUUID()
-        val userName = "jpt-$uuid"
-        val payload = """
-            {
-                "name": "$userName",
-                "password": "${target.password}",
-                "emailAddress": "$userName@testing.com",
-                "displayName": "New JPT VU $userName"
-            }
-            """.trimIndent()
-        val requestBody = RequestBody.create(MediaType.parse("application/json"), payload)
+
         val credential = Credentials.basic(target.userName, target.password)
-        val request = Request.Builder()
+        val reqBuilder = Request.Builder()
             .url(target.webApplication.resolve("rest/api/2/user").toString())
             .header("Authorization", credential)
+        val userName = IdempotentAction("create user via REST") {
+            createUser(reqBuilder, target.password)
+        }.retry(5, ExponentialBackoff(Duration.ofSeconds(1)))
+
+        return User(name = userName, password = target.password)
+    }
+
+    private fun createUser(reqBuilder: Request.Builder, password: String) : String {
+        val givenNamesUsedInThisAttempt = nameGenerator.getNumberOfGivenNamesToGenerate()
+        val (userName, payload) = createUserPayload(password, givenNamesUsedInThisAttempt)
+        val requestBody = RequestBody.create(MediaType.parse(APPLICATION_JSON.toString()), payload)
+        val request = reqBuilder
             .post(requestBody)
             .build()
+
         time("create user via REST") {
             try {
                 httpClient.newCall(request).execute()
@@ -48,17 +56,39 @@ class RestUserGenerator(readTimeout: Duration) : UserGenerator {
                 throw e
             }
         }.use { response ->
-            if (response.code() == 201) {
+            if (response.code() == HttpURLConnection.HTTP_CREATED) {
+                nameGenerator.onSuccess()
                 logger.info("Created a new user $userName")
+                return userName
             } else {
-                val msg = "Failed to create a new user $userName:" +
+                val possibleReason = if (response.code() == HttpURLConnection.HTTP_BAD_REQUEST) {
+                    nameGenerator.onCollision(givenNamesUsedInThisAttempt)
+                    "assuming name collision"
+                } else {
+                    "unknown reason"
+                }
+                val msg = "Failed to create a new user $userName, " +
+                    possibleReason + ": " +
                     " response code ${response.code()}," +
-                    " response body ${response.body()?.string()}"
+                    " response body [${response.body()?.string()}]" + response.headers()
 
                 logger.error(msg)
                 throw Exception(msg)
             }
         }
-        return User(name = userName, password = target.password)
+    }
+
+    private fun createUserPayload(password: String, givenNamesCnt: Int): Pair<String, String> {
+        val displayName = nameGenerator.pickRandomUnique(givenNamesCnt)
+        val userName = UserNameUtils.toUserName(displayName)
+        val payload = """
+                {
+                    "name": "$userName",
+                    "password": "$password",
+                    "emailAddress": "$userName@jpt-testing.com",
+                    "displayName": "$displayName"
+                }
+                """.trimIndent()
+        return Pair(userName, payload)
     }
 }