+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/virtual-users/compare/release-3.10.0...master
 
+### Added
+- Generate realistic user names in `RestUserGenerator`.
+
 ## [3.10.0] - 2019-08-02
 [3.10.0]: https://github.com/atlassian/virtual-users/compare/release-3.9.1...release-3.10.0